LayoutTests/fast/forms/search/search-field-background-color.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
<style>
    input { background: red; }
</style>
</head>

<body>
<p>The following search field should have red background.</p>
<input type="search" id="test">

<script>
if (window.testRunner)
    testRunner.dumpAsText(true);

    var test = document.getElementById("test");
    var color = 'rgb(255, 0, 0)';
    shouldBe("getComputedStyle(test, null).getPropertyValue('background-color')", "color");
</script>

</body>
</html>
