LayoutTests/fast/forms/search/search-field-background-color-expected.txt

PASS getComputedStyle(test, null).getPropertyValue('background-color') is color
The following search field should have red background.
